+++ src/api2/managers/documents-manager.spec.js
@@ -1,17 +1,17 @@
 import sinon from 'sinon'
 
-import { DocumentUploader } from './document-uploader'
-import { ApiCaller } from './api-caller'
+import { DocumentsManager } from './documents-manager'
+import { ApiCaller } from '../api-caller'
 
-describe('file-uploader unit test', () => {
+describe('documents-manager unit test', () => {
   let sandbox
-  let documentUploader
+  let documentsManager
 
   beforeEach(() => {
     sandbox = sinon.createSandbox()
 
     const apiCaller = ApiCaller.getInstance('https://api.com')
-    documentUploader = new DocumentUploader({ apiCaller })
+    documentsManager = new DocumentsManager({ apiCaller })
   })
 
   afterEach(() => {
@@ -21,19 +21,19 @@ describe('file-uploader unit test', () => {
   describe('method', () => {
     describe('uploadDocument', () => {
       it('calls proper methods and returns file storage key', async () => {
-        sandbox.stub(documentUploader, '_createDocumentAnchorConfig').resolves({
+        sandbox.stub(documentsManager, '_createDocumentAnchorConfig').resolves({
           id: 'id',
           url: 'https://storage.com',
           type: 'type',
           key: 'doc-key',
           somePolicy: 'value'
         })
-        sandbox.stub(documentUploader, '_createFileFormData').returns({
+        sandbox.stub(documentsManager, '_createFileFormData').returns({
           key: 'value'
         })
-        sandbox.stub(documentUploader, '_postFormData').resolves()
+        sandbox.stub(documentsManager, '_postFormData').resolves()
 
-        const result = await documentUploader.uploadDocument({
+        const result = await documentsManager.uploadDocument({
           type: 'type',
           mimeType: 'mime-type',
           file: { name: 'file' },
@@ -41,13 +41,13 @@ describe('file-uploader unit test', () => {
         })
 
         expect(result).to.equal('doc-key')
-        expect(documentUploader._createDocumentAnchorConfig)
+        expect(documentsManager._createDocumentAnchorConfig)
           .calledOnceWithExactly({
             type: 'type',
             mimeType: 'mime-type',
             accountId: 'SOME_ACCOUNT_ID'
           })
-        expect(documentUploader._createFileFormData)
+        expect(documentsManager._createFileFormData)
           .calledOnceWithExactly({
             file: { name: 'file' },
             policy: {
@@ -55,7 +55,7 @@ describe('file-uploader unit test', () => {
               somePolicy: 'value'
             }
           })
-        expect(documentUploader._postFormData)
+        expect(documentsManager._postFormData)
           .calledOnceWithExactly({ key: 'value' })
       }
       )
@@ -63,17 +63,17 @@ describe('file-uploader unit test', () => {
 
     describe('_createDocumentAnchorConfig', () => {
       it('calls apiCaller.postWithSignature method with provided params', async () => {
-        sandbox.stub(documentUploader._apiCaller, 'wallet')
+        sandbox.stub(documentsManager._apiCaller, 'wallet')
           .get(_ => ({ accountId: 'SOME_ACCOUNT_ID' }))
-        sandbox.stub(documentUploader._apiCaller, 'postWithSignature')
+        sandbox.stub(documentsManager._apiCaller, 'postWithSignature')
           .resolves({ data: { key: 'doc-key' } })
 
-        const result = await documentUploader._createDocumentAnchorConfig({
+        const result = await documentsManager._createDocumentAnchorConfig({
           type: 'doc-type',
           mimeType: 'mime-type'
         })
 
-        expect(documentUploader._apiCaller.postWithSignature)
+        expect(documentsManager._apiCaller.postWithSignature)
           .to.have.been.calledOnceWithExactly(
             '/documents',
             {
@@ -94,12 +94,12 @@ describe('file-uploader unit test', () => {
 
     describe('_postFormData', () => {
       it('posts file form data', async () => {
-        sandbox.stub(documentUploader._axios, 'post').resolves()
-        documentUploader.useStorageURL('https://storage.com')
+        sandbox.stub(documentsManager._axios, 'post').resolves()
+        documentsManager.useStorageURL('https://storage.com')
 
-        await documentUploader._postFormData({ key: 'value' })
+        await documentsManager._postFormData({ key: 'value' })
 
-        expect(documentUploader._axios.post).calledOnceWithExactly(
+        expect(documentsManager._axios.post).calledOnceWithExactly(
           'https://storage.com',
           { key: 'value' },
           { headers: { 'Content-Type': 'multipart/form-data' } }