+++ src/index.js
@@ -11,9 +11,10 @@ export { ApiResponse } from './api'
 
 export {
   ApiCaller,
-  DocumentUploader,
-  WalletsManager,
+  DocumentsManager,
   FactorsManager,
+  SignersManager,
+  WalletsManager,
   JsonapiResponse
 } from './api2'
 export { KeyServerCaller } from './api2/key-server-caller'